TST: Rearrange files
