+++ databroker/_mds/metadataclient/tests/test_mds.py
@@ -2,7 +2,7 @@ import time as ttime
 import uuid
 import pytest
 from ....headersource.mongo_core import NoEventDescriptors, NoRunStart
-from ....tests.mds_tests.test_mds import (
+from ....tests.test_mds import (
     test_bad_bulk_insert_event_data,
     test_bad_bulk_insert_event_timestamp,
     test_bad_event_desc,
